Cleanup and separation of serialisation tests
